File pond

+++ src/lib/input/BadgeWithDel.svelte
@@ -3,11 +3,14 @@
 	import { createEventDispatcher } from 'svelte';
 	const dispatch = createEventDispatcher();
 	export let text;
+	export let withDeleteButton = true;
 </script>
 
 <Badge pill color="light" class="kfk-tag text-primary border border-primary">
 	{text}
-	<a href={'#'} on:click|preventDefault|stopPropagation={() => dispatch('delete', 'delete this')}>
-		<Icon name="x" />
-	</a>
+	{#if withDeleteButton}
+		<a href={'#'} on:click|preventDefault|stopPropagation={() => dispatch('delete', 'delete this')}>
+			<Icon name="x" />
+		</a>
+	{/if}
 </Badge>
